contracts dividendsFor and votingFor include decimals

+++ contracts/contracts/layer_2/snapshots/SnapshotsStorageWrapper_2.sol
@@ -247,13 +247,13 @@ abstract contract SnapshotsStorageWrapper_2 is
 
     function _decimalsAtSnapshot(
         uint256 _snapshotID
-    ) internal view returns (uint256 decimals_) {
+    ) internal view returns (uint8 decimals_) {
         (bool snapshotted, uint256 value) = _valueAt(
             _snapshotID,
             _snapshotStorage_2().decimals
         );
 
-        return snapshotted ? value : _decimals();
+        return snapshotted ? uint8(value) : _decimals();
     }
 
     function _updateAccountSnapshot(
